Fix string overrun in ddm.c
There was a possible off-by-one due to the string terminator
if the generated program was big enough, and the realloc wasn't
correctly reallocing the space.

So I removed this part of the code and reverted to my old-school
dynstr from my personal C library that should do the trick.

With this fix, the python interpreter is no longer crashing when
calling into the DDM C code.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>