Preliminary implementation of rebinding with ddm optimization
This commit introduces most of the glue between the ddm C code and
the python metasimulation stuff.

I did not have much time to work on the rebinding of actors itself.
Indeed, it's violating an assertion right now. Unfortunately my time
is over for a while.

The sim_from_trace_ddm.py is essentially a copy of sim_from_trace.py.
Considering that we will need a "sim_from_trace_random.py" script later
on, it would be better to have a hierarchy of python objects that
implement what should be done upon a TIME_WINDOW event (ie: collect
statistics, assign using ddm, assign using random, etc) and instantiate
one in sim_from_trace.py depending on what experiments we want to run.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>